+++ e2e/sedge/generate_test.go
@@ -1,0 +1,182 @@
+/*
+Copyright 2022 Nethermind
+
+Licensed under the Apache License, Version 2.0 (the "License");
+you may not use this file except in compliance with the License.
+You may obtain a copy of the License at
+
+	http://www.apache.org/licenses/LICENSE-2.0
+
+Unless required by applicable law or agreed to in writing, software
+distributed under the License is distributed on an "AS IS" BASIS,
+WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+See the License for the specific language governing permissions and
+limitations under the License.
+*/
+package e2e
+
+import (
+	"path/filepath"
+	"strings"
+	"testing"
+
+	"github.com/NethermindEth/sedge/internal/lido/contracts"
+	"github.com/NethermindEth/sedge/internal/lido/contracts/mevboostrelaylist"
+	"github.com/joho/godotenv"
+	"github.com/stretchr/testify/assert"
+
+	base "github.com/NethermindEth/sedge/e2e"
+)
+
+func TestE2E_Generate_FullNode_GoerliNotSupported(t *testing.T) {
+	// Test context
+	var (
+		runErr error
+	)
+	// Build test case
+	e2eTest := newE2ESedgeTestCase(
+		t,
+		// Arrange
+		nil,
+		// Act
+		func(t *testing.T, binaryPath, dataDirPath string) {
+			runErr = base.RunSedge(t, binaryPath, "generate", "full-node", "--network", "goerli")
+		},
+		// Assert
+		func(t *testing.T, dataDirPath string) {
+			assert.Error(t, runErr, "generate command should fail without arguments")
+		},
+	)
+	// Run test case
+	e2eTest.run()
+}
+
+func TestE2E_Generate_FullNode_Lido_GnosisNotSupported(t *testing.T) {
+	// Test context
+	var (
+		runErr error
+	)
+	// Build test case
+	e2eTest := newE2ESedgeTestCase(
+		t,
+		// Arrange
+		nil,
+		// Act
+		func(t *testing.T, binaryPath string, dataDirPath string) {
+			runErr = base.RunSedge(t, binaryPath, "generate", "--lido", "full-node", "--network", "gnosis")
+		},
+		// Assert
+		func(t *testing.T, dataDirPath string) {
+			assert.Error(t, runErr, "generate command should fail")
+		},
+	)
+	// Run test case
+	e2eTest.run()
+}
+
+func TestE2E_Generate_FullNode_Lido_SepoliaNotSupported(t *testing.T) {
+	// Test context
+	var (
+		runErr error
+	)
+	// Build test case
+	e2eTest := newE2ESedgeTestCase(
+		t,
+		// Arrange
+		nil,
+		// Act
+		func(t *testing.T, binaryPath string, dataDirPath string) {
+			runErr = base.RunSedge(t, binaryPath, "generate", "--lido", "full-node", "--network", "sepolia")
+		},
+		// Assert
+		func(t *testing.T, dataDirPath string) {
+			// Sepolia not supported Lido with MEV-Boost
+			assert.Error(t, runErr, "generate command should fail")
+		},
+	)
+	// Run test case
+	e2eTest.run()
+}
+
+func TestE2E_Generate_FullNode_Lido_Sepolia_NoMEV(t *testing.T) {
+	// Test context
+	var (
+		runErr error
+	)
+	// Build test case
+	e2eTest := newE2ESedgeTestCase(
+		t,
+		// Arrange
+		nil,
+		// Act
+		func(t *testing.T, binaryPath string, dataDirPath string) {
+			runErr = base.RunSedge(t, binaryPath, "generate", "--lido", "full-node", "--network", "sepolia", "--no-mev-boost")
+		},
+		// Assert
+		func(t *testing.T, dataDirPath string) {
+			assert.NoError(t, runErr, "generate command should succeed with the given arguments")
+			generateDataFilePath := filepath.Join(dataDirPath, ".env")
+			assert.FileExists(t, generateDataFilePath, ".env file should be created")
+
+			// Read .env file
+			envMap, err := godotenv.Read(generateDataFilePath)
+			assert.NoError(t, err, "should be able to read .env file")
+
+			// Read FEE_RECIPIENT value
+			feeRecipient, exists := envMap["FEE_RECIPIENT"]
+			assert.True(t, exists, "FEE_RECIPIENT should exist in .env file")
+			expectedFeeRecipient, ok := contracts.FeeRecipient("sepolia")
+			assert.True(t, ok, "FeeRecipient should be found")
+			assert.Equal(t, expectedFeeRecipient, feeRecipient, "FEE_RECIPIENT value should match expected value")
+
+			// Read RELAY_URLS value
+			_, exists = envMap["RELAY_URLS"]
+			assert.False(t, exists, "RELAY_URLS shouldn't exist in .env file")
+		},
+	)
+	// Run test case
+	e2eTest.run()
+}
+
+func TestE2E_Generate_FullNode_Lido_Mainnet(t *testing.T) {
+	// Test context
+	var (
+		runErr error
+	)
+	// Build test case
+	e2eTest := newE2ESedgeTestCase(
+		t,
+		// Arrange
+		nil,
+		// Act
+		func(t *testing.T, binaryPath string, dataDirPath string) {
+			runErr = base.RunSedge(t, binaryPath, "generate", "--lido", "full-node", "--network", "mainnet")
+		},
+		// Assert
+		func(t *testing.T, dataDirPath string) {
+			assert.NoError(t, runErr, "generate command should succeed")
+			generateDataFilePath := filepath.Join(dataDirPath, ".env")
+			assert.FileExists(t, generateDataFilePath, ".env file should be created")
+
+			// Read .env file
+			envMap, err := godotenv.Read(generateDataFilePath)
+			assert.NoError(t, err, "should be able to read .env file")
+
+			// Read FEE_RECIPIENT value
+			feeRecipient, exists := envMap["FEE_RECIPIENT"]
+			assert.True(t, exists, "FEE_RECIPIENT should exist in .env file")
+			expectedFeeRecipient, ok := contracts.FeeRecipient("mainnet")
+			assert.True(t, ok, "FeeRecipient should be found")
+			assert.Equal(t, expectedFeeRecipient, feeRecipient, "FEE_RECIPIENT value should match expected value")
+
+			// Read RELAY_URLS value
+			relayURLs, exists := envMap["RELAY_URLS"]
+			assert.True(t, exists, "RELAY_URLS should exist in .env file")
+			relayURLsList := strings.Split(relayURLs, ",")
+			expectedRelayURLs, _ := mevboostrelaylist.RelaysURI("mainnet")
+			assert.Equal(t, expectedRelayURLs, relayURLsList, "RELAY_URLS value should match expected value")
+		},
+	)
+	// Run test case
+	e2eTest.run()
+}